RavenDB-18852 Boost query fix. BoostingMultiTermMatch is not supporting boosting with And.
