+++ src/Corax/IndexSearcher/IndexSearcher.Boost.cs
@@ -10,24 +10,7 @@ public partial class IndexSearcher
     {
         return BoostingMatch.WithConstant(this, match, constant);
     }
-
-    public BoostingMatch Boost<TInner, TScoreFunction>(TInner match, TScoreFunction scoreFunction)
-        where TInner : IQueryMatch
-        where TScoreFunction : IQueryScoreFunction
-    {
-        if (typeof(TScoreFunction) == typeof(TermFrequencyScoreFunction))
-        {
-            return BoostingMatch.WithTermFrequency(this, match, (TermFrequencyScoreFunction)(object)scoreFunction);
-        }
-        else if (typeof(TScoreFunction) == typeof(ConstantScoreFunction))
-        {
-            return BoostingMatch.WithConstant(this, match, (ConstantScoreFunction)(object)scoreFunction);
-        }
-        else
-        {
-            throw new NotSupportedException($"Boosting with the score function {typeof(TScoreFunction)} is not supported.");
-        }
-    }
+    
     
     public BoostingMatch Boost<TInner>(TInner match, IQueryScoreFunction scoreFunction)
         where TInner : IQueryMatch
